requirements.txt
Django==4.0.2
django-crum==0.7.9
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
flake8==4.0.1
numpy==1.22.2